Lots of changes in this commit. First of all the Java inspired identifiers (com.zen.sections for example) are no longer used but instead packages, plugins and themes should simply specify a lowercase name that's unique. Second of all a lot of internals have been changed and cleaned up to make it easier for me to develop them in the future.
Modules such as Zen::Language and Zen::Plugin have been modified so that they can be
included into the global namespace without having to include useless class methods that
shouldn't be used on their own anyway.

A new interesting feature is Zen::Validation, a basic validation "framework" for classes 
that's used to ensure the attributes of plugins, themes and packages are set correctly. A
basic example of using this module would look like the following:

    class SomeClass
      include Zen::Validation

      attr_accessor :name
      attr_accessor :age

      def validate
        validates_presence(:name)
        validates_format(:name, /[0-9]+/)
      end
    end

Currently the following still has to be done:

* Update the API documentation for Zen::Plugin, Zen::Package and Zen::Theme.
* Add the ability to have controller specific access rules.
* Replace all Thor tasks by Rake tasks now that Thor is no longer really required.