Lots of changes in this commit. First of all Thor has been removed from the project. It was originally introduced because at the time I needed to re-use tasks in a situation where Rake wasn't available. Because Thor tasks are just classes it was much easier to use Thor as a solution at the time. Because I recently fixed the problem (not going into the details) Thor was no longer required and thus I decided to bring back Rake. On top of all that I find that Rake has a better codebase and feels less like a Rails approach to running tasks.
Some other features are a new executable file that's written using Commander and allows me 
to more easily extend it's features. Next to this I also merged the String extension from 
the ruby-extensions gem with Zen itself. It's rather useless to depend on such a small Gem 
when you can just include the single file into the project.

In short, the following happened:

* Thor was replaced with Rake
* The zen executable was re-written using Commander
* General code cleanups
* Began working on a new ACL system that no longer requires Traits.