Added the ability to load a global Javascript or CSS file from the application's public directory. What this means is that if you were to create the file public/admin/js/global.js or public/admin/css/global.css it will always be loaded in the backend. This can be extremely useful if you want to override a certain style or add a set of features without having to write a package or plugin for it. Combine this with Yepnope and CSS @import declarations and you can quickly restyle the entire backend.
Last not but least, if you happen to have multiple public directories
set in Ramaze.options.publics only the first existing file will be used.
This means that if you have directory A and B that both contain a
global.js file only the one from A will be used.