Initially I didn't want to split this commit up into multiple "smaller" ones but to ensure I don't lose any changes I decided to do it anyway.
Quite a lot has changed in this commit so far. The biggest addition is that users
are now able to define their own field types. This means that if you want to
allow a client to add Email fields you can do so. If your field type is using
a method that does not shit with Zen itself you can add it quite easily but it
will require some knowledge of the BlueForm helper.

So how does this work? Well, managing custom field types is broken up in two
different parts. There's the database and the corresponding CRUD actions and
there's the module CustomFields::BlueFormParameters (what a name eh?). The first
bit is used to store the custom field types and the methods they use (these
methods should be defined in the BlueFormParameters module). When building the
custom fields in the section entries form Zen uses the data from the database
and passes this to the BlueFormParameters module. This module generates the
required parameters for each custom field which are then passed to
Ramaze::Helper::BlueForm.

The advantage of this approach is that in the event of a method name being
changed or extra methods being added to BlueForm it's quite easy to allow people
to use these in Zen as well. On top of that it's possible to easily add your own
methods.

Besides this rather big feature I've also changed several smaller things:

* The controller methods edit() and new() now use the form.xtml view directly.
* visual_editor (and all variations) has been renamed to "text_editor".
* tables no longer require a class of "table" for certain CSS styles and
  Javascript features.
* General code cleanups.

The following still has to be done:

* "css_class" and "css_id" should be renamed to "html_class" and "html_id"
  respectively.
* The sections package should be modified so that it uses the new custom field
  type feature (currently it's broken).
* HTML tables should be sortable and should have alternative colors for odd
  rows.
* Expand all the Bacon specs and document them similar to the ones used for the
  custom fields package.
* Write specs for the custom field types feature.