Removed Ramaze::Helper::Common and cleaned up the users package.
The helper Ramaze::Helper::Common was only used because of the anchor_to()
method. Now that that method is no longer used (in favor of calling .a() on a
controller's class) I decided to remove it.

The users package has been modified so that it follows the same coding style as
all the other packages.