Moved all the multi select dropdowns over to checkboxes.
<select> elements with multiple="multiple" are annoying to use as they require
you to hold an extra key depending on your OS just to select multiple values. On
top of that they can take quite a bit of vertical space if you have a dropdown
with multiple possible values.

Along with this change comes a small fix that solves the issue of not being
able to add categories to a section entry.