Moved all the error classes into a single file and added a hook system.
The hook system (Zen::Hook) can be used to execute blocks of code whenever a
certain action is triggered. An example of this could be a hook that's executed
whenever a new comment is submitted by a user. Currently Zen itself does not use
any hooks but I'll add a few in the following commits.