User registration and language fixes.
When registering a new user, either by using the web interface or by directly
using the model, the password is now set/required properly. Previously you would
be able to register a user without a password under certain conditions which
would cause BCrypt to get angry. This patch modifies the User model so that it
*only* updates the password for an existing user if the password isn't nil or
empty.

This patch also comes with a small fix related to a language string that was
plural when it was supposed to be singular.

Signed-off-by: YorickPeterse <yorickpeterse@gmail.com>