Javascript localization and spec layouts.
Rather than dumping a big lump of JSON directly into the body Zen now uses
Zen::Controller::Language to generate a JSON object of the current language.
This controller is set up in such a way that it behaves like a normal Javascript
file.

Another fix introduced in this commit is that layouts are now working when
running tests. This required some minor modifications to some tests but should
make the overal process of testing the code easier.

Signed-off-by: YorickPeterse <yorickpeterse@gmail.com>