HTML5 form validation.
All the forms in the admin panel now use a set of attributes that should trigger
a browser's form validation system if that particular browser has such a feature
available. For example, Chrome will show a warning telling you that a certain
field is required if it has an ``required`` attribute set. This feature isn't
something I'm heavily relying on, it's more of a "nice to have".

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>