Specify user statuses with IDs.
Previously the status of a user was stored as a string of which the value could
be "open" or "closed". Instead of these string based values the Users package
now uses IDs to associate a status with a user. The following statuses are
available:

* active (allows login)
* closed
* unconfirmed

The last status will be used for users that have registered but haven't
confirmed their account yet.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>