Gave the settings API its much needed love.
Setting values are now serialized. This makes it possible for settings to have
multiple values set (which is useful for checkboxes/select fields with multiple
possible values). While implementing this I noticed that the settings system
never supported checkboxes and a few other fields. You could add them but they
would never be displayed in the settings form. This has been fixed as well.

The improved API also no longer needs the database columns for the default
value(s) of a setting, its type, etc. I'm not sure why I added those to the
database to begin with but since they weren't used at all I decided to remove
these.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>