Store translations in blocks.
This approach makes it possible for language files to be re-loaded if certain
keys are missing. This can be extremely useful if you're using an external cache
such as memcached as it means you don't have to restart your application in the
event of loss of data in the cache, instead it will simply be re-created.

This commit also marks the last step in migrating all language files and
documenting them. Rough benchmarks showed that the total amount of memory being
used in :live mode has decreased a bit less than I hoped for. The good thing is
that if you use an external cache (memcached!) the memory usage won't grow
either. Previously it would grow for every YAML file being parsed.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>