Default template groups for themes.
Template groups and sections are no longer bound to each other. Previously the
setting "default section" would only be used to figure out what template group
to render. Because template groups don't really have anything to do with
sections (they can be used for them but it isn't always needed) I decided to
remove this.

Instead of this developers can now specify a default template group to use when
creating their themes (this is set to "default") as following:

    Zen::Theme.add do |t|
      t.default_template_group = 'foobar'
    end

The setting "default_section" is no longer being used and can be removed from
your settings table.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>