Various improvements related to editing users.
Users are now able to always edit their own account regardless of whether or not
they have the permission "edit_user". While working on this I discovered that in
this case users were able to add themselves to existing user groups or change
their status. To prevent this from happening I added two new permissions
required for this:

* edit_user_status
* assign_user_group

This fixes #41.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>