Work around SQLite3 foreign key issues.
After 2 days of debugging with Jeremy evans we managed to pinpoint the issue
with SQLite3 and migrations when using Sequel 3.31.0. It turns out that the use
of foreign keys combined with the way you have to alter tables in SQLite3 causes
the database to become corrupted.

To work around this you have to turn foreign keys off (PRAGMA foreign_keys =
off), run the migrations and then turn it back on. To make this a bit easier I
wrote a small wrapper for Sequel::Migrator called Zen::Migrator. This module
takes care of the SQLite3 nonsense as well as logging messages (making the Rake
tasks a bit cleaner).

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>