Performance improvements for render_menu().
The performance of the method Ramaze::Helper::MenuFrontend#render_menu() has
been greatly improved by using a more efficient way of retrieving child items of
a given menu item.

Before this commit the MenuItem model would use Sequel's tree plugin for
retrieving child elements. The problem with this plugin is that it will execute
a query for each menu item to retrieve all child items. This would in many cases
result in a large amount of queries being generated. Especially with menus with
a large hierarchy this could easily result in 20, maybe 30 queries (or more
depending on the amount of items).

To solve this issue I wrote a new method: Menus::Model::Menu#menu_item_tree().
This method is capable of building a hierarchy of menu items for a given menu
using only two queries: one to retrieve the actual menu (it's an instance method
so you need to retrieve a menu yourself first) and one query to retrieve all the
items of that menu. The tree building process itself is done in memory.

The performance increase of this change is quite drastic. The average siege
benchmark should show an increase of about 10 transactions per second for the
Zen website when using a relative simple menu structure. However, based on the
complexity of your menus the increase might be even bigger.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>