Fixed various LRU related issues
The LRU cache would in certain cases complain about not being able to dump
singleton objects. The workaround for this is to not directly store model
instances in the flash data but instead only store their values. Another 
advantage of this change is that it will most likely reduce the memory usage a
bit, although I haven't benchmarked this.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>