Load core packages when connecting to a DB.
This is the first step of making it easier to load third-party packages as well
as creating settings without having to do so after calling Zen.start and
manually migrating the settings (which would also mean you can't use custom
assets).

Whenever a database connection is created the various model related settings
(such as the plugins to use globally) as well as the core packages are loaded.
To make this possible the various translation strings for Sequel are "lazy
loaded" using lambdas.

What this means is that instead of the following:

    require 'zen'

    require __DIR__('config/database')

    Zen.start

    require 'your/custom/package'

You can now do this:

    require 'zen'

    require __DIR__('config/database')
    require 'your/custom/package'

    Zen.start

Currently you'd still have to manually migrate settings (if you register any
custom ones) since there's no guaranteed order for events to be executed in.
I'll address this issue in the following commits.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>