Simplified the loading of Gems.
Rubygems (apparently) takes care of ensuring that whatever version you have
defined in your Gemspec for a certain Gem is actually used, even if there are
newer versions of this gem installed. This feature makes the use of both Bundler
and Ramaze.setup somewhat useless.

When it comes to development of Zen itself people will just have to ensure that
the Gems they have in their Gemset are the correct versions. Using Bundler for
this is not an option because of the following reasons:

* It assumes your Gemfile is in $PWD unless you set an environment variable,
  this in turn would more than likely make it impossible to load other Gemfiles
  (e.g. by apps using a local copy of Zen) without first resetting this
  variable.
* Bundler adds an average of 5-6 Megabytes of RAM. This was benchmarked by
  running the Zen website in live mode using a single Thin worker and the
  following command:

      $ siege -c 10 -t 10s -b http://localhost:7000/ > /dev/null

  Measuring memory usage was done with the following command:

      $ ps aux | grep thin | awk '{print $6/1024}'

* Bundler is bloated. Capistran recipes? Deployment strategies? Why would I need
  all that when all I care is about loading specific Gem versions and easily
  installing them?

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>